+++ src/main/java/de/tum/in/www1/artemis/web/rest/dto/TutorLeaderboardDTO.java
@@ -20,11 +20,11 @@ public class TutorLeaderboardDTO {
 
     private long numberOfTutorMoreFeedbackRequests;
 
-    private Long points;
+    private double points;
 
     public TutorLeaderboardDTO(long userId, String name, long numberOfAssessments, long numberOfAcceptedComplaints, long numberOfTutorComplaints,
             long numberOfNotAnsweredMoreFeedbackRequests, long numberOfComplaintResponses, long numberOfAnsweredMoreFeedbackRequests, long numberOfTutorMoreFeedbackRequests,
-            Long points) {
+            double points) {
         this.userId = userId;
         this.name = name;
         this.numberOfAssessments = numberOfAssessments;
@@ -109,11 +109,11 @@ public class TutorLeaderboardDTO {
         this.numberOfTutorMoreFeedbackRequests = numberOfTutorMoreFeedbackRequests;
     }
 
-    public Long getPoints() {
+    public double getPoints() {
         return points;
     }
 
-    public void setPoints(Long points) {
+    public void setPoints(double points) {
         this.points = points;
     }
 }
